updating end points for MU authentication

+++ routes/notes.js
@@ -2,21 +2,27 @@
 
 const express = require('express');
 const router = express.Router();
-
+const passport = require('passport');
 const mongoose = require('mongoose');
-
 const Note = require('../models/note');
+
+
+router.use('/', passport.authenticate('jwt', { session: false, failWithError: true }));
+
 
 /* ========== GET/READ ALL ITEMS ========== */
 router.get('/', (req, res, next) => {
   const { searchTerm, folderId, tagId } = req.query;
+  const userId = req.user.id;
 
-  let filter = {};
+  let filter ={ userId};
+
 
   if (searchTerm) {
-    // filter.title = { $regex: searchTerm };
-    filter.$or = [{ 'title': { $regex: searchTerm } }, { 'content': { $regex: searchTerm } }];
+    const re = new RegExp(searchTerm, 'i');
+    filter.title = { $regex: re };
   }
+
 
   if (folderId) {
     filter.folderId = folderId;
@@ -26,7 +32,9 @@ router.get('/', (req, res, next) => {
     filter.tags = tagId;
   }
 
-  Note.find(filter)
+  console.log(filter);
+
+  Note.find( filter)
     .populate('tags')
     .sort({ 'updatedAt': 'desc' })
     .then(results => {
@@ -40,6 +48,9 @@ router.get('/', (req, res, next) => {
 /* ========== GET/READ A SINGLE ITEM ========== */
 router.get('/:id', (req, res, next) => {
   const { id } = req.params;
+  const userId = req.user.id;
+
+  console.log(userId);
 
   if (!mongoose.Types.ObjectId.isValid(id)) {
     const err = new Error('The `id` is not valid');
@@ -47,7 +58,7 @@ router.get('/:id', (req, res, next) => {
     return next(err);
   }
 
-  Note.findById(id)
+  Note.findOne({ _id: id, userId })
     .populate('tags')
     .then(result => {
       if (result) {
@@ -64,6 +75,7 @@ router.get('/:id', (req, res, next) => {
 /* ========== POST/CREATE AN ITEM ========== */
 router.post('/', (req, res, next) => {
   const { title, content, folderId, tags = [] } = req.body;
+  const userId = req.user.id;
 
   /***** Never trust users - validate input *****/
   if (!title) {
@@ -88,7 +100,7 @@ router.post('/', (req, res, next) => {
     });
   }
 
-  Note.create({ title, content, folderId, tags })
+  Note.create({ title, userId, content, folderId, tags })
     .then(result => {
       res
         .location(`${req.originalUrl}/${result.id}`)
@@ -104,7 +116,9 @@ router.post('/', (req, res, next) => {
 router.put('/:id', (req, res, next) => {
   const { id } = req.params;
   const { title, content, folderId, tags = [] } = req.body;
+  const {userId} = req.user.id;
 
+  console.log(id);
   /***** Never trust users - validate input *****/
   if (!mongoose.Types.ObjectId.isValid(id)) {
     const err = new Error('The `id` is not valid');
@@ -134,7 +148,7 @@ router.put('/:id', (req, res, next) => {
     });
   }
 
-  Note.findByIdAndUpdate(id, { title, content, folderId, tags }, { new: true })
+  Note.findByIdAndUpdate( {_id:id, userId}, { title, content, folderId, tags }, { new: true })
     .then(result => {
       if (result) {
         res.json(result);
@@ -150,8 +164,9 @@ router.put('/:id', (req, res, next) => {
 /* ========== DELETE/REMOVE A SINGLE ITEM ========== */
 router.delete('/:id', (req, res, next) => {
   const { id } = req.params;
+  const {userId} = req.user.id;
 
-  Note.findByIdAndRemove(id)
+  Note.findByIdAndRemove({_id:id, userId})
     .then(() => {
       res.status(204).end();
     })
